contact page styling is done.
