+++ styles/Contact.module.scss
@@ -1,30 +1,22 @@
-.modalContent {
-  font-size: 3rem;
-  letter-spacing: 2px;
-  background-color: rgb(0, 247, 255);
-  border: 1px solid #888;
-  position: relative;
-  top: 30%;
+.containter {
   text-align: center;
+  background-color: aqua;
+  height: 100vh;
 
-  button {
-    border: none;
-    font-size: 3rem;
-    letter-spacing: 2px;
-    padding: 5px 25px;
-    border-radius: 20px;
-    margin-bottom: 20px;
-    background-color: yellow;
-    font-family: "Bangers", cursive;
-
-    &:hover {
-      color: white;
-      background-color: red;
-    }
+  h2 {
+    font-size: 4rem;
+    margin: 20px 0;
   }
-}
 
-.container {
-  margin-left: auto;
-  background-color: blue;
+  p {
+    font-size: 2rem;
+    margin: 20px 0;
+  }
+
+  h2,
+  p {
+    padding: 0;
+
+    letter-spacing: 2px;
+  }
 }
